+++ app/Http/Controllers/BidController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\BidRequest;
 use App\Models\Bid;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BidController extends Controller
 {
@@ -13,6 +14,8 @@ class BidController extends Controller
      */
     public function store(BidRequest $request)
     {
+        Gate::authorize('create', Bid::class);
+
         Bid::create([
             'ad_id' => $request->ad_id,
             'user_id' => Auth::id(),
@@ -27,6 +30,8 @@ class BidController extends Controller
      */
     public function destroy(Bid $bid)
     {
+        Gate::authorize('delete', $bid);
+
         $bid->delete();
 
         return redirect()->route('ads.show', $bid->ad_id);
